fix: rename `fetch_examples.py` -> `dataset_helper.py`
