+++ pyrit/datasets/adv_bench_dataset.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import List, Literal, Optional
 
 from pyrit.common.path import DATASETS_PATH
-from pyrit.datasets.fetch_examples import fetch_examples
+from pyrit.datasets.dataset_helper import fetch_examples
 from pyrit.models import SeedPromptDataset
 from pyrit.models.seed_prompt import SeedPrompt
