+++ pyrit/datasets/aya_redteaming_dataset.py
@@ -4,7 +4,7 @@
 from pathlib import Path
 from typing import List, Literal, Optional
 
-from pyrit.datasets.fetch_examples import fetch_examples
+from pyrit.datasets.dataset_helper import fetch_examples
 from pyrit.models import SeedPromptDataset
 from pyrit.models.seed_prompt import SeedPrompt
