+++ pyrit/datasets/decoding_trust_stereotypes_dataset.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import List, Literal, Optional
 
 from pyrit.common.path import DATASETS_PATH
-from pyrit.datasets.fetch_examples import FILE_TYPE_HANDLERS, fetch_examples
+from pyrit.datasets.dataset_helper import FILE_TYPE_HANDLERS, fetch_examples
 from pyrit.models import SeedPromptDataset
 from pyrit.models.seed_prompt import SeedPrompt
