+++ pyrit/datasets/harmbench_dataset.py
@@ -4,7 +4,7 @@
 from pathlib import Path
 from typing import Literal, Optional
 
-from pyrit.datasets.fetch_examples import FILE_TYPE_HANDLERS, fetch_examples
+from pyrit.datasets.dataset_helper import FILE_TYPE_HANDLERS, fetch_examples
 from pyrit.models import SeedPromptDataset
 from pyrit.models.seed_prompt import SeedPrompt
